refactor: 메시지 전송 클래스 구현을 위해 기존 클래스 및 파일 이름 변경
