+++ src/app/application/group/command/disablePortfolio/DisablePortfolioCommandHandler.spec.ts
@@ -5,7 +5,7 @@ import { DisablePortfolioCommand } from '@khlug/app/application/group/command/di
 import { DisablePortfolioCommandHandler } from '@khlug/app/application/group/command/disablePortfolio/DisablePortfolioCommandHandler';
 
 import {
-  ISlackSender,
+  IMessageSender,
   SlackSender,
 } from '@khlug/app/domain/adapter/ISlackSender';
 import {
@@ -24,7 +24,7 @@ import { PointGrantService } from '@khlug/app/domain/user/service/PointGrantServ
 
 import { DomainFixture } from '@khlug/__test__/fixtures';
 import { GroupFixture } from '@khlug/__test__/fixtures/GroupFixture';
-import { Message } from '@khlug/constant/message';
+import { Message } from '@khlug/constant/error';
 import { Point } from '@khlug/constant/point';
 
 describe('DisablePortfolioCommandHandler', () => {
@@ -33,7 +33,7 @@ describe('DisablePortfolioCommandHandler', () => {
   let groupMemberRepository: jest.Mocked<IGroupMemberRepository>;
   let groupLogger: jest.Mocked<IGroupLogger>;
   let pointGrantService: jest.Mocked<PointGrantService>;
-  let slackSender: jest.Mocked<ISlackSender>;
+  let slackSender: jest.Mocked<IMessageSender>;
 
   beforeEach(async () => {
     advanceTo(new Date());
