+++ src/app/application/infraBlue/query/getDoorLockPassword/GetDoorLockPasswordQueryHandler.ts
@@ -8,7 +8,7 @@ import { GetDoorLockPasswordQueryResult } from '@khlug/app/application/infraBlue
 
 import { Cache, CacheId } from '@khlug/app/domain/cache/model/Cache';
 
-import { Message } from '@khlug/constant/message';
+import { Message } from '@khlug/constant/error';
 
 @QueryHandler(GetDoorLockPasswordQuery)
 export class GetDoorLockPasswordQueryHandler
