+++ src/app/application/user/command/createDiscordIntegration/CreateDiscordIntegrationCommandHandler.spec.ts
@@ -19,7 +19,7 @@ import {
 } from '@khlug/app/domain/discord/IDiscordIntegrationRepository';
 
 import { DiscordIntegrationFixture } from '@khlug/__test__/fixtures/DiscordIntegrationFixture';
-import { Message } from '@khlug/constant/message';
+import { Message } from '@khlug/constant/error';
 
 describe('CreateDiscordIntegrationCommandHandler', () => {
   let handler: CreateDiscordIntegrationCommandHandler;
