+++ src/app/application/user/command/updateUser/UpdateUserCommandHandler.ts
@@ -21,7 +21,7 @@ import {
   UserRepository,
 } from '@khlug/app/domain/user/IUserRepository';
 
-import { Message } from '@khlug/constant/message';
+import { Message } from '@khlug/constant/error';
 
 @CommandHandler(UpdateUserCommand)
 export class UpdateUserCommandHandler
