+++ src/app/application/user/eventHandler/UserProfileUpdatedHandler.ts
@@ -2,10 +2,10 @@ import { Inject } from '@nestjs/common';
 import { EventsHandler, IEventHandler } from '@nestjs/cqrs';
 
 import {
-  ISlackSender,
+  IMessageSender,
   SlackSender,
 } from '@khlug/app/domain/adapter/ISlackSender';
-import { SlackMessageCategory } from '@khlug/app/domain/message/model/constant';
+import { MessageCategory } from '@khlug/constant/message';
 import { UserProfileUpdated } from '@khlug/app/domain/user/event/UserProfileUpdated';
 
 @EventsHandler(UserProfileUpdated)
@@ -14,7 +14,7 @@ export class UserProfileUpdatedHandler
 {
   constructor(
     @Inject(SlackSender)
-    private readonly slackSender: ISlackSender,
+    private readonly slackSender: IMessageSender,
   ) {}
 
   handle(event: UserProfileUpdated) {
@@ -23,7 +23,7 @@ export class UserProfileUpdatedHandler
     this.slackSender.send({
       targetUserId: user.id,
       message: '회원 정보를 수정했습니다.',
-      category: SlackMessageCategory.USER_DATA,
+      category: MessageCategory.USER_DATA,
     });
   }
 }
